Extract Logger into its own package (#2165)
