+++ packages/core/logger/.gitignore
@@ -1,0 +1,1 @@
+lib
