+++ packages/core/logger/src/.babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [["@babel/preset-env", {
+    "targets": {
+      "node": "6"
+    }
+  }]],
+  "plugins": ["@babel/plugin-transform-runtime"]
+}
