+++ packages/core/logger/src/emoji.js
