+++ packages/core/logger/test/mocha.opts
@@ -1,0 +1,2 @@
+--require @parcel/babel-register
+--exit
