+++ packages/core/parcel-bundler/src/utils/generateCertificate.js
@@ -2,7 +2,7 @@ const forge = require('node-forge');
 const fs = require('fs');
 const mkdirp = require('mkdirp');
 const path = require('path');
-const logger = require('../Logger');
+const logger = require('@parcel/logger');
 
 function generateCertificate(options = {}) {
   const privateKeyPath = path.join(options.cacheDir, 'private.pem');
