+++ packages/core/parcel-bundler/src/utils/installPackage.js
@@ -2,7 +2,7 @@ const config = require('./config');
 const promisify = require('./promisify');
 const resolve = promisify(require('resolve'));
 const commandExists = require('command-exists');
-const logger = require('../Logger');
+const logger = require('@parcel/logger');
 const pipeSpawn = require('./pipeSpawn');
 const PromiseQueue = require('./PromiseQueue');
 const path = require('path');
