+++ packages/core/parcel-bundler/src/visitors/fs.js
@@ -2,7 +2,7 @@ const t = require('@babel/types');
 const Path = require('path');
 const fs = require('fs');
 const template = require('@babel/template').default;
-const logger = require('../Logger');
+const logger = require('@parcel/logger');
 
 const bufferTemplate = template('Buffer(CONTENT, ENC)');
